Replacing GetTor, TorBEL, and TorCheck with TAILS, TorStatus, and the Metrics Portal on the projects page. The reason is that these are substantial user-facing projects, and hence far more likely of interest to visitors (besides, the replaced projects were mostly just pointing to trac entry stubs).
I've double checked with Sebastian (the owner of the only semi-active of those projects) to make sure the change was alright with him, and they all have entries in the volunteer section so potential contributors can still find them. I also double checked on #tails that this entry is fine with them. The project listing on the frontpage is being left unchanged.