Storing copies of past gsoc proposals
Suggested by Roger in case the originals vanish. These are pdf printoffs
with the exception of hbock's DNSEL rewrite app, which didn't translate well.
This was a simple page, so just copying it.

I'm also replacing a simple txt copy of jvoisin's metadata toolkit proposal
with a much better looking pdf of his melange app (restricted to just the
content via firebug).