Adding Tails, Orbot, TorStatus, and arm to the site front page
This adds these four projects to the "Our Projects" section of the Tor front
page. Unlike the previous projects that we had there (Torbutton, Tor Check,
Vidalia, and TBB) these aren't part of the TBB ecosystem and hance not easily
discovered.

Having all eight projects feels cluttered to me and I'd suggest dropping Check
and TBB since the user will be downloading and using these reguardless of a
presence on the front page, but this is Andrew's call.