Flagging GeoIP project as being ineligible for GSoC
We previously had a note saying that this was a no-go for GSoC but it had been
commented out. If this was git then I'd use 'git blame' to figure out why, but
not worth the effort to try to get this information from svn.

Suggested revision by Robert Ransom.