Dropping the remaining orbot projects
Nathan hasn't responded about GSoC mentorship, and his last email indicated
that it was unlikely. Jake gave a blanket 'yea, I can mentor any of those
projects' which is nice, but doens't fill me with much confidence that we have
a primary mentor for anything Orbot related.

I'm on the fence about if we should remove them or flag as simply not being
eligible for GSoC and mostly decided to go this way due to lack of
communication. If Nathan replies later then great.