More accurately describe start time (I hope).
Also, if people want to avoid mad Google juice for their name
and salary in our tax returns, this is neither legally nor
technically required of them as contractors. (Which is what
the job post claims to be for).

Finally, if Steve Jobs gets to demand Insanely Great, I get
to demand Insanely Creative.