Few minor tweaks for the volunteer table
As I understand it Aaron has adopted the torctl/torflow codebases. They also
haven't been touched in months so noting both.

Ravi's still working on stem and Sathyanarayanan is working on Onionoo's python
counterpart (it should probably be listed as a separate project, but he hasn't
provided an entry for it yet).