Only listing project ideas with mentors
As theatened on tor-assistants@ I'm trimming our volunteer page's task listing
to only include things with an engaged mentor.

The listing has grown stale which is confusing to newcomers. Some projects are
done, and others reflect plans that are a year out of date. My devious scheme
was that if I required some effort from mentors to keep task ideas around then
we'd only get buy in for those that were really still relevant. Looks like it
worked - we're down to three.