Dropping JTor from the project table
Ok, I'm adopting the following policy:

"If a project never made it out of alpha *and* hasn't been touched in over two
years then we should stop listing it."

Thandy and TorBEL are getting close to this threshold, but JTor has been
inactive for almost three years. It's confusing for us to keep showing
prospective volunteers such a long dead project.