Dropping stem's PathSupport project from the page
Last year I proposed a stem counterpart for PathSupport as a project for GSoC.
This did not go very well since we really haven't a clue *what* we want from
such a project. Since then this has not changed and there's little reason to
leave it on the page (even commented out).