Adding a project to expand stem's tests to more dedicatedly test tor
This is a project that I have been batting around with Nick for a bit. It
doesn't include the 'grand vision' he wants for covering inter-network
components, but it would provide dedicated tests for a very substantial portion
of Tor's codebase.

While I think a more complicated network simulator and testing suite would
yield greater benefits, I also think it would be *vastly* harder to make and
hence quite likely to never happen. Presently our automated testing of Tor is
pretty poor and there's no need to let the perfect be the enemy of the good on
this.