Dropping commented out project ideas
Why are we still bundling dozens of commented out tasks with the page? If we
want to add these back in later then we should... well, add them back in later.
That's what SCMs are for (even if svn sucks at it). :P

This is slowing down page loads, and most ideas are getting to be pretty old
anyway.