Raising the JTor activity to moderate
Bruce and velope requested for it to be raised. I'm tempted to say 'heavy'
based on the number of commits, but prior to March there was no activity at all
for two years so I'm gonna hold off on doing that until we see if this is
maintained or a flash in the pan.