Replacing vidalia with stem on the front page
I've been a little worried about stem's discoverability for a while now. Over
this last month I've been investing a lot of time into making stem's site
(https://stem.torproject.org/) beginner friendly for newcomers that want to
develop against tor, but that's all for not if they don't even know that it
exists.

Vidalia is both inactive and no longer used on its own (it's either used as
part of the Tor Browser or not at all), hence there is little reason that
visitors to our site would be interested in knowing about it.

Replacing Vidalia with Stem on the project grid of the front page. I'm also
replacing stem's logo with one of a slightly higher quality. I asked armadev on
irc if he had any objections to this but no reply (my guess is that he doesn't
really care much, though it might also be that he's too distracted with the
obfs2 blocking discussion and forgot my question).